[*] Featured Products module: admin area updated
git-svn-id: svn://svn.crtdev.local/repo/xlite/main@2279 194c3f1c-803d-db11-9c9e-000e0c8563f3

+++ test/classes/XLite/Module/FeaturedProducts/Model/Category.php
@@ -30,7 +30,7 @@
  * XLite_Module_FeaturedProducts_Model_Category 
  * 
  * @package    XLite
- * @subpackage ____sub_package____
+ * @subpackage Model
  * @since      3.0.0
  */
 class XLite_Module_FeaturedProducts_Model_Category extends XLite_Model_Category implements XLite_Base_IDecorator
@@ -77,7 +77,7 @@ class XLite_Module_FeaturedProducts_Model_Category extends XLite_Model_Category 
     /**
      * Add specified products to the featured products list
      * 
-     * @param mixed $products ____param_comment____
+     * @param array $products Array of XLite_Model_Product objects
      *  
      * @return void
      * @access public
@@ -86,20 +86,22 @@ class XLite_Module_FeaturedProducts_Model_Category extends XLite_Model_Category 
      */
 	public function addFeaturedProducts($products)
 	{
-		for ($i=0; $i<count($products); $i++) {
-			$fp = new XLite_Module_FeaturedProducts_Model_FeaturedProduct();
-			$fp->set("category_id", $this->get("category_id"));
-			$fp->set("product_id", $products[$i]->get("product_id"));
-			if (!$fp->isExists()) {
-				$fp->create();
-			}
-		}
+        if (is_array($products)) {
+            foreach ($products as $product) {
+    			$fp = new XLite_Module_FeaturedProducts_Model_FeaturedProduct();
+    			$fp->set("category_id", $this->get("category_id"));
+       			$fp->set("product_id", $product->get("product_id"));
+        		if (!$fp->isExists()) {
+		    		$fp->create();
+    			}
+    		}
+        }
 	}
 
     /**
      * Delete specified products from the featured products list
      * 
-     * @param mixed $products ____param_comment____
+     * @param array $products Array of XLite_Model_Product objects
      *  
      * @return void
      * @access public
@@ -108,12 +110,14 @@ class XLite_Module_FeaturedProducts_Model_Category extends XLite_Model_Category 
      */
 	public function deleteFeaturedProducts($products)
 	{
-		for ($i=0; $i<count($products); $i++) {
-			$fp = new XLite_Module_FeaturedProducts_Model_FeaturedProduct();
-			$fp->set("category_id", $this->get("category_id"));
-			$fp->set("product_id", $products[$i]->get("product_id"));
-			$fp->delete();
-		}	
+        if (is_array($products)) {
+            foreach ($products as $product) {
+    			$fp = new XLite_Module_FeaturedProducts_Model_FeaturedProduct();
+    			$fp->set("category_id", $this->get('category_id'));
+    			$fp->set("product_id", $product->get('product_id'));
+    			$fp->delete();
+    		}
+        }
 	}
 
     /**
